+++ djangocms_stories/feeds.py
@@ -38,11 +38,9 @@ class LatestEntriesFeed(Feed):
         return _("Blog articles on %(site_name)s") % {"site_name": Site.objects.get_current().name}
 
     def items(self, obj=None):
-        return (
-            Post.objects.prefetch_related("postcontent_set")
-            .filter(app_config__namespace=self.namespace, include_in_rss=True)
-            .order_by("-date_published")[: self.feed_items_number]
-        )
+        return Post.objects.prefetch_related("postcontent_set").filter(
+            app_config__namespace=self.namespace, include_in_rss=True
+        )[: self.feed_items_number][: self.feed_items_number]
 
     def item_title(self, item):
         return mark_safe(item.safe_translation_getter("title"))