+++ CHANGELOG.rst
@@ -2,6 +2,25 @@
 Changelog
 =========
 
+0.9.3 (2026-06-28)
+------------------
+
+* feat: Add sorted autocomplete to related posts plugin by @fsbraun in https://github.com/django-cms/djangocms-stories/pull/107
+* feat: Replace autosuggest by autocomplete select2 widget by @fsbraun in https://github.com/django-cms/djangocms-stories/pull/108
+* feat: Add new locales by @fsbraun in https://github.com/django-cms/djangocms-stories/pull/109
+
+0.9.2 (2026-06-12)
+------------------
+
+* feat: Declare root template to django CMS 5.1 by @fsbraun in https://github.com/django-cms/djangocms-stories/pull/102
+* fix: Avoid posts with date_published set to fall behind all without by @MacLake in https://github.com/django-cms/djangocms-stories/pull/101
+
+
+0.9.1 (2026-05-14)
+------------------
+
+* fix: Prefetch detection used wrong attribute by @fsbraun in https://github.com/django-cms/djangocms-stories/pull/103
+
 0.9.0 (2026-05-01)
 ------------------